+++ lib/multi_split_view.dart
@@ -3,4 +3,5 @@ library multi_split_view;
 export 'src/multi_split_view.dart';
 export 'src/theme_widget.dart';
 export 'src/theme_data.dart';
+export 'src/area.dart';
 export 'src/divider_painter.dart';