+++ lib/src/multi_split_view.dart
@@ -2,7 +2,9 @@ import 'dart:collection';
 import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:multi_split_view/src/area.dart';
 import 'package:multi_split_view/src/divider_widget.dart';
+import 'package:multi_split_view/src/sizes_cache.dart';
 import 'package:multi_split_view/src/theme_data.dart';
 import 'package:multi_split_view/src/theme_widget.dart';
 
@@ -13,95 +15,124 @@ import 'package:multi_split_view/src/theme_widget.dart';
 class MultiSplitViewController extends ChangeNotifier {
   static const double _higherPrecision = 1.0000000000001;
 
-  MultiSplitViewController._(this._weights);
-
   /// Creates an [MultiSplitViewController].
   ///
   /// The sum of the [weights] cannot exceed 1.
-  factory MultiSplitViewController({List<double>? weights}) {
-    return MultiSplitViewController._(
-        weights != null ? List.from(weights) : []);
+  factory MultiSplitViewController({List<Area>? areas}) {
+    return MultiSplitViewController._(areas != null ? List.from(areas) : []);
   }
 
-  List<double> _weights;
-  UnmodifiableListView<double> get weights => UnmodifiableListView(_weights);
+  MultiSplitViewController._(this._areas);
 
-  set weights(List<double> weights) {
-    _weights = List.from(weights);
+  List<Area> _areas;
+  UnmodifiableListView<Area> get areas => UnmodifiableListView(_areas);
+
+  set areas(List<Area> areas) {
+    _areas = List.from(areas);
     notifyListeners();
   }
 
-  /// Gets the weight of a given widget.
-  double getWeight(int index) {
-    return _weights[index];
+  /// Gets the area of a given widget index.
+  Area getArea(int index) {
+    return _areas[index];
   }
 
-  void _setWeight(int index, double value) {
-    if (value < 0) {
-      throw Exception('Weight needs to be positive: $value');
-    }
-    _weights[index] = value;
+  /// Sum of all weights.
+  double _weightSum() {
+    double sum = 0;
+    _areas.forEach((area) {
+      sum += area.weight ?? 0;
+    });
+    return sum;
   }
 
   /// Adjusts the weights according to the number of children.
   /// New children will receive a percentage of current children.
   /// Excluded children will distribute their weights to the existing ones.
   @visibleForTesting
-  void validateAndAdjust(int childrenCount) {
+  void fixWeights(
+      {required int childrenCount,
+      required double fullSize,
+      required double dividerThickness}) {
     childrenCount = math.max(childrenCount, 0);
 
-    double weightSum = 0;
-    for (int i = 0; i < _weights.length; i++) {
-      double weight = _weights[i];
-      if (weight <= 0) {
-        throw Exception('Weight needs to be positive: $weight');
+    final double totalDividerSize = (childrenCount - 1) * dividerThickness;
+    final double availableSize = fullSize - totalDividerSize;
+
+    int nullCount = 0;
+    for (int i = 0; i < _areas.length; i++) {
+      Area area = _areas[i];
+      if (area.size != null) {
+        _areas[i] = area.copyWithNewWeight(weight: area.size! / availableSize);
       }
-      weightSum += weight;
+      if (area.weight == null) {
+        nullCount++;
+      }
     }
 
+    double weightSum = _weightSum();
+
+    // fill null values
+    if (nullCount > 0) {
+      double r = 0;
+      if (weightSum < MultiSplitViewController._higherPrecision) {
+        r = (1 - weightSum) / nullCount;
+      }
+      for (int i = 0; i < _areas.length; i++) {
+        Area area = _areas[i];
+        if (area.weight == null) {
+          _areas[i] = area.copyWithNewWeight(weight: r);
+        }
+      }
+      weightSum = _weightSum();
+    }
+
+    // removing over weight...
     if (weightSum > MultiSplitViewController._higherPrecision) {
-      throw Exception('The sum of the weights cannot exceed 1: $weightSum');
+      final over = weightSum - 1;
+      double r = over / weightSum;
+      for (int i = 0; i < _areas.length; i++) {
+        Area area = _areas[i];
+        _areas[i] =
+            area.copyWithNewWeight(weight: area.weight! - (area.weight! * r));
+      }
     }
 
-    if (_weights.length == childrenCount) {
+    if (_areas.length == childrenCount) {
       _fillWeightsEqually(childrenCount, weightSum);
       return;
-    } else if (_weights.length < childrenCount) {
+    } else if (_areas.length < childrenCount) {
       // children has been added.
-      int addedChildrenCount = childrenCount - _weights.length;
+      int addedChildrenCount = childrenCount - _areas.length;
       double newWeight = 0;
       if (weightSum < 1) {
         newWeight = (1 - weightSum) / addedChildrenCount;
       } else {
-        for (int i = 0; i < _weights.length; i++) {
-          double r = _weights[i] / childrenCount;
-          _weights[i] -= r;
+        for (int i = 0; i < _areas.length; i++) {
+          Area area = _areas[i];
+          double r = area.weight! / childrenCount;
+          _areas[i] = area.copyWithNewWeight(weight: area.weight! - r);
           newWeight += r / addedChildrenCount;
         }
       }
       for (int i = 0; i < addedChildrenCount; i++) {
-        _weights.add(newWeight);
+        _areas.add(Area(weight: newWeight));
       }
     } else {
       // children has been removed.
       double removedWeight = 0;
-      while (_weights.length > childrenCount) {
-        removedWeight += _weights.removeLast();
+      while (_areas.length > childrenCount) {
+        removedWeight += _areas.removeLast().weight!;
       }
-      if (_weights.isNotEmpty) {
-        double w = removedWeight / _weights.length;
-        for (int i = 0; i < _weights.length; i++) {
-          _weights[i] += w;
+      if (_areas.isNotEmpty) {
+        double w = removedWeight / _areas.length;
+        for (int i = 0; i < _areas.length; i++) {
+          Area area = _areas[i];
+          _areas[i] = area.copyWithNewWeight(weight: area.weight! + w);
         }
       }
     }
-
-    weightSum = 0;
-    _weights.forEach((weight) {
-      weightSum += weight;
-    });
-
-    _fillWeightsEqually(childrenCount, weightSum);
+    _fillWeightsEqually(childrenCount, _weightSum());
   }
 
   /// Fills equally the missing weights
@@ -110,8 +141,9 @@ class MultiSplitViewController extends ChangeNotifier {
       double availableWeight = 1 - weightSum;
       if (availableWeight > 0) {
         double w = availableWeight / childrenCount;
-        for (int i = 0; i < _weights.length; i++) {
-          _weights[i] += w;
+        for (int i = 0; i < _areas.length; i++) {
+          Area area = _areas[i];
+          _areas[i] = area.copyWithNewWeight(weight: area.weight! + w);
         }
       }
     }
@@ -126,9 +158,6 @@ class MultiSplitViewController extends ChangeNotifier {
 /// A widget to provides horizontal or vertical multiple split view.
 class MultiSplitView extends StatefulWidget {
   static const Axis defaultAxis = Axis.horizontal;
-  static const double defaultMinimalWeight = .01;
-  static const double minimalWidgetWeightLowerLimit = 0.01;
-  static const double minimalWidgetWeightUpperLimit = 0.9;
 
   /// Creates an [MultiSplitView].
   ///
@@ -142,44 +171,19 @@ class MultiSplitView extends StatefulWidget {
       this.axis = MultiSplitView.defaultAxis,
       required this.children,
       this.controller,
-      this.globalMinimalWeight,
-      List<double>? minimalWeights,
       this.dividerBuilder,
-      this.globalMinimalSize,
-      List<double>? minimalSizes,
       this.onSizeChange,
       this.resizable = true,
       this.antiAliasingWorkaround = true,
-      List<double>? initialWeights})
-      : this.initialWeights =
-            initialWeights != null ? List.from(initialWeights) : null,
-        this.minimalWeights =
-            minimalWeights != null ? List.from(minimalWeights) : null,
-        this.minimalSizes =
-            minimalSizes != null ? List.from(minimalSizes) : null,
-        super(key: key) {
-    if (globalMinimalWeight != null &&
-        (globalMinimalWeight! < MultiSplitView.minimalWidgetWeightLowerLimit ||
-            globalMinimalWeight! >
-                MultiSplitView.minimalWidgetWeightUpperLimit)) {
-      throw Exception(
-          'The global minimum weight must be between ${MultiSplitView.minimalWidgetWeightLowerLimit} and ${MultiSplitView.minimalWidgetWeightUpperLimit}.');
-    }
-    if (minimalWeights != null) {
-      minimalWeights.forEach((value) {
-        if (value < MultiSplitView.minimalWidgetWeightLowerLimit ||
-            value > MultiSplitView.minimalWidgetWeightUpperLimit) {
-          throw Exception(
-              'The minimum weight must be between ${MultiSplitView.minimalWidgetWeightLowerLimit} and ${MultiSplitView.minimalWidgetWeightUpperLimit}.');
-        }
-      });
-    }
-  }
+      List<Area>? initialAreas})
+      : this.initialAreas =
+            initialAreas != null ? List.from(initialAreas) : null,
+        super(key: key);
 
   final Axis axis;
   final List<Widget> children;
   final MultiSplitViewController? controller;
-  final List<double>? initialWeights;
+  final List<Area>? initialAreas;
 
   /// Defines a builder of dividers. Overrides the default divider
   /// created by the theme.
@@ -187,29 +191,6 @@ class MultiSplitView extends StatefulWidget {
 
   /// Indicates whether it is resizable. The default value is [TRUE].
   final bool resizable;
-
-  /// Defines the global minimal weight for all children. The default value is defined by [MultiSplitView.defaultMinimalWeight] constant.
-  /// It will be used if [minimalWeights] has not been set.
-  final double? globalMinimalWeight;
-
-  /// Defines the minimal weight for each child.
-  final List<double>? minimalWeights;
-
-  /// Defines the global minimal size in pixels for all children.
-  /// It will be used if [minimalSizes] has not been set.
-  /// The size will be converted into weight and will respect the limit
-  /// defined by the [MultiSplitView.defaultMinimalWeight] constant,
-  /// allowing all children to be visible.
-  /// The value [zero] is ignored and indicates that no size has been set.
-  final double? globalMinimalSize;
-
-  /// Defines the minimal size in pixels for each child.
-  /// It will be used if [globalMinimalWeight] has not been set.
-  /// The size will be converted into weight and will respect the limit
-  /// defined by the [MultiSplitView.defaultMinimalWeight] constant,
-  /// allowing all children to be visible.
-  /// The value [zero] is ignored and indicates that no size has been set.
-  final List<double>? minimalSizes;
 
   /// Function to listen any children size change.
   final OnSizeChange? onSizeChange;
@@ -224,19 +205,18 @@ class MultiSplitView extends StatefulWidget {
 /// State for [MultiSplitView]
 class _MultiSplitViewState extends State<MultiSplitView> {
   late MultiSplitViewController _controller;
-  double _initialDragPos = 0;
-  double _initialChild1Weight = 0;
-  double _initialChild2Weight = 0;
-  double _initialChild1Size = 0;
+  _InitialDrag? _initialDrag;
+
   int? _draggingDividerIndex;
   int? _hoverDividerIndex;
+  SizesCache? _sizesCache;
 
   @override
   void initState() {
     super.initState();
     _controller = widget.controller != null
         ? widget.controller!
-        : MultiSplitViewController(weights: widget.initialWeights);
+        : MultiSplitViewController(areas: widget.initialAreas);
     _stateHashCodeValidation();
     _controller._stateHashCode = hashCode;
     _controller.addListener(_rebuild);
@@ -249,7 +229,9 @@ class _MultiSplitViewState extends State<MultiSplitView> {
   }
 
   void _rebuild() {
-    setState(() {});
+    setState(() {
+      _sizesCache = null;
+    });
   }
 
   @override
@@ -263,13 +245,13 @@ class _MultiSplitViewState extends State<MultiSplitView> {
     super.didUpdateWidget(oldWidget);
 
     if (widget.controller != _controller) {
-      List<double> weights = _controller.weights;
+      List<Area> areas = _controller._areas;
       _controller._stateHashCode = null;
       _controller.removeListener(_rebuild);
 
       _controller = widget.controller != null
           ? widget.controller!
-          : MultiSplitViewController(weights: weights);
+          : MultiSplitViewController(areas: areas);
       _stateHashCodeValidation();
       _controller._stateHashCode = hashCode;
       _controller.addListener(_rebuild);
@@ -294,25 +276,46 @@ class _MultiSplitViewState extends State<MultiSplitView> {
           (widget.children.length - 1) * themeData.dividerThickness;
 
       return LayoutBuilder(builder: (context, constraints) {
-        _controller.validateAndAdjust(widget.children.length);
         List<Widget> children = [];
         if (widget.axis == Axis.horizontal) {
-          double availableArea = constraints.maxWidth - totalDividerSize;
+          _controller.fixWeights(
+              childrenCount: widget.children.length,
+              fullSize: constraints.maxWidth,
+              dividerThickness: themeData.dividerThickness);
+          if (_sizesCache == null ||
+              _sizesCache!.childrenCount != widget.children.length ||
+              _sizesCache!.fullSize != constraints.maxWidth) {
+            _sizesCache = SizesCache(
+                areas: _controller._areas,
+                fullSize: constraints.maxWidth,
+                dividerThickness: themeData.dividerThickness);
+          }
           _populateHorizontalChildren(
               context: context,
               constraints: constraints,
               totalDividerSize: totalDividerSize,
               children: children,
-              availableArea: availableArea,
+              fullSize: constraints.maxWidth,
               themeData: themeData);
         } else {
-          double availableArea = constraints.maxHeight - totalDividerSize;
+          _controller.fixWeights(
+              childrenCount: widget.children.length,
+              fullSize: constraints.maxHeight,
+              dividerThickness: themeData.dividerThickness);
+          if (_sizesCache == null ||
+              _sizesCache!.childrenCount != widget.children.length ||
+              _sizesCache!.fullSize != constraints.maxHeight) {
+            _sizesCache = SizesCache(
+                areas: _controller._areas,
+                fullSize: constraints.maxHeight,
+                dividerThickness: themeData.dividerThickness);
+          }
           _populateVerticalChildren(
               context: context,
               constraints: constraints,
               totalDividerSize: totalDividerSize,
               children: children,
-              availableArea: availableArea,
+              fullSize: constraints.maxHeight,
               themeData: themeData);
         }
 
@@ -320,42 +323,6 @@ class _MultiSplitViewState extends State<MultiSplitView> {
       });
     }
     return Container();
-  }
-
-  /// Calculates the minimum weight. Used when the pixel size has been set.
-  double _minimalWeight(
-      {required int childIndex, required double availableArea}) {
-    if (widget.minimalWeights != null &&
-        childIndex < widget.minimalWeights!.length) {
-      return widget.minimalWeights![childIndex];
-    } else if (widget.globalMinimalWeight != null) {
-      return widget.globalMinimalWeight!;
-    }
-
-    if (widget.minimalSizes != null &&
-        childIndex < widget.minimalSizes!.length) {
-      double size = math.max(0, widget.minimalSizes![childIndex]);
-      if (size > 0) {
-        return _sizeToWeight(size: size, availableArea: availableArea);
-      }
-    }
-
-    if (widget.globalMinimalSize != null) {
-      double size = math.max(0, widget.globalMinimalSize!);
-      if (size > 0) {
-        return _sizeToWeight(size: size, availableArea: availableArea);
-      }
-    }
-    return MultiSplitView.defaultMinimalWeight;
-  }
-
-  /// Calculates a weight given a size in pixels.
-  double _sizeToWeight({required double size, required double availableArea}) {
-    double minimalWeight = size / availableArea;
-    minimalWeight =
-        math.max(minimalWeight, MultiSplitView.defaultMinimalWeight);
-    minimalWeight = math.min(minimalWeight, 1 / widget.children.length);
-    return minimalWeight;
   }
 
   /// Updates the hover divider index.
@@ -375,31 +342,22 @@ class _MultiSplitViewState extends State<MultiSplitView> {
       required BoxConstraints constraints,
       required double totalDividerSize,
       required List<Widget> children,
-      required double availableArea,
+      required double fullSize,
       required MultiSplitViewThemeData themeData}) {
-    double totalChildrenSize = constraints.maxWidth - totalDividerSize;
-    double totalRemainingWeight = 1;
     _DistanceFrom childDistance = _DistanceFrom();
     for (int childIndex = 0;
         childIndex < widget.children.length;
         childIndex++) {
       bool highlighted = (_draggingDividerIndex == childIndex ||
           (_draggingDividerIndex == null && _hoverDividerIndex == childIndex));
-      double childWeight = _controller.getWeight(childIndex);
-      totalRemainingWeight -= childWeight;
 
-      childDistance.right = _distanceOnTheOppositeSide(
-          childIndex: childIndex,
-          totalChildrenSize: totalChildrenSize,
-          totalRemainingWeight: totalRemainingWeight,
-          themeData: themeData);
+      final double childSize = _sizesCache!.sizes[childIndex];
+      childDistance.right = fullSize - childSize - childDistance.left;
 
       children.add(_buildPositioned(
           distance: childDistance, child: widget.children[childIndex]));
 
       if (childIndex < widget.children.length - 1) {
-        double childSize = totalChildrenSize * childWeight;
-
         _DistanceFrom dividerDistance = _DistanceFrom();
         dividerDistance.left = childDistance.left + childSize;
         dividerDistance.right =
@@ -428,21 +386,18 @@ class _MultiSplitViewState extends State<MultiSplitView> {
                   _draggingDividerIndex = childIndex;
                 });
                 final pos = _position(context, detail.globalPosition);
-                _updateInitialValues(childIndex, pos.dx, totalChildrenSize);
+                _updateInitialDrag(childIndex, pos.dx);
               },
-              onHorizontalDragEnd: (detail) {
-                setState(() {
-                  _draggingDividerIndex = null;
-                });
-              },
+              onHorizontalDragEnd: (detail) => _onDragEnd(),
               onHorizontalDragUpdate: (detail) {
                 final pos = _position(context, detail.globalPosition);
-                double diffX = pos.dx - _initialDragPos;
+                double diffX = pos.dx - _initialDrag!.initialDragPos;
 
                 _updateDifferentWeights(
                     childIndex: childIndex,
                     diffPos: diffX,
-                    availableArea: availableArea);
+                    fullSize: fullSize,
+                    pos: pos.dx);
               },
               child: dividerWidget);
           dividerWidget = _mouseRegion(
@@ -464,24 +419,17 @@ class _MultiSplitViewState extends State<MultiSplitView> {
       required BoxConstraints constraints,
       required double totalDividerSize,
       required List<Widget> children,
-      required double availableArea,
+      required double fullSize,
       required MultiSplitViewThemeData themeData}) {
-    double totalChildrenSize = constraints.maxHeight - totalDividerSize;
-    double totalRemainingWeight = 1;
     _DistanceFrom childDistance = _DistanceFrom();
     for (int childIndex = 0;
         childIndex < widget.children.length;
         childIndex++) {
       bool highlighted = (_draggingDividerIndex == childIndex ||
           (_draggingDividerIndex == null && _hoverDividerIndex == childIndex));
-      double childWeight = _controller.getWeight(childIndex);
-      totalRemainingWeight -= childWeight;
 
-      childDistance.bottom = _distanceOnTheOppositeSide(
-          childIndex: childIndex,
-          totalChildrenSize: totalChildrenSize,
-          totalRemainingWeight: totalRemainingWeight,
-          themeData: themeData);
+      final double childSize = _sizesCache!.sizes[childIndex];
+      childDistance.bottom = fullSize - childSize - childDistance.top;
 
       children.add(_buildPositioned(
           distance: childDistance,
@@ -489,8 +437,6 @@ class _MultiSplitViewState extends State<MultiSplitView> {
           last: childIndex == widget.children.length - 1));
 
       if (childIndex < widget.children.length - 1) {
-        double childSize = totalChildrenSize * childWeight;
-
         _DistanceFrom dividerDistance = _DistanceFrom();
         dividerDistance.top = childDistance.top + childSize;
         dividerDistance.bottom =
@@ -519,20 +465,17 @@ class _MultiSplitViewState extends State<MultiSplitView> {
                   _draggingDividerIndex = childIndex;
                 });
                 final pos = _position(context, detail.globalPosition);
-                _updateInitialValues(childIndex, pos.dy, totalChildrenSize);
+                _updateInitialDrag(childIndex, pos.dy);
               },
-              onVerticalDragEnd: (detail) {
-                setState(() {
-                  _draggingDividerIndex = null;
-                });
-              },
+              onVerticalDragEnd: (detail) => _onDragEnd(),
               onVerticalDragUpdate: (detail) {
                 final pos = _position(context, detail.globalPosition);
-                double diffY = pos.dy - _initialDragPos;
+                double diffY = pos.dy - _initialDrag!.initialDragPos;
                 _updateDifferentWeights(
                     childIndex: childIndex,
                     diffPos: diffY,
-                    availableArea: availableArea);
+                    fullSize: fullSize,
+                    pos: pos.dy);
               },
               child: dividerWidget);
           dividerWidget = _mouseRegion(
@@ -546,6 +489,19 @@ class _MultiSplitViewState extends State<MultiSplitView> {
         childDistance.top = dividerDistance.top + themeData.dividerThickness;
       }
     }
+  }
+
+  void _onDragEnd() {
+    for (int i = 0; i < _controller._areas.length; i++) {
+      Area area = _controller._areas[i];
+      double size = _sizesCache!.sizes[i];
+      _controller._areas[i] =
+          area.copyWithNewWeight(weight: size / _sizesCache!.childrenSize);
+    }
+
+    setState(() {
+      _draggingDividerIndex = null;
+    });
   }
 
   /// Wraps the divider widget with a [MouseRegion].
@@ -565,48 +521,118 @@ class _MultiSplitViewState extends State<MultiSplitView> {
         child: dividerWidget);
   }
 
-  void _updateInitialValues(
-      int childIndex, double pos, double totalChildrenSize) {
-    _initialDragPos = pos;
-    _initialChild1Weight = _controller.getWeight(childIndex);
-    _initialChild2Weight = _controller.getWeight(childIndex + 1);
-    _initialChild1Size = totalChildrenSize * _initialChild1Weight;
+  void _updateInitialDrag(int childIndex, double initialDragPos) {
+    final double initialChild1Size = _sizesCache!.sizes[childIndex];
+    final double initialChild2Size = _sizesCache!.sizes[childIndex + 1];
+    final double minimalChild1Size = _sizesCache!.minimalSizes[childIndex];
+    final double minimalChild2Size = _sizesCache!.minimalSizes[childIndex + 1];
+    final double sumMinimals = minimalChild1Size + minimalChild2Size;
+    final double sumSizes = initialChild1Size + initialChild2Size;
+
+    double posLimitStart = 0;
+    double posLimitEnd = 0;
+    double child1Start = 0;
+    double child2End = 0;
+    for (int i = 0; i <= childIndex; i++) {
+      if (i < childIndex) {
+        child1Start += _sizesCache!.sizes[i];
+        child1Start += _sizesCache!.dividerThickness;
+        child2End += _sizesCache!.sizes[i];
+        child2End += _sizesCache!.dividerThickness;
+        posLimitStart += _sizesCache!.sizes[i];
+        posLimitStart += _sizesCache!.dividerThickness;
+        posLimitEnd += _sizesCache!.sizes[i];
+        posLimitEnd += _sizesCache!.dividerThickness;
+      } else if (i == childIndex) {
+        posLimitStart += _sizesCache!.minimalSizes[i];
+        posLimitEnd += _sizesCache!.sizes[i];
+        posLimitEnd += _sizesCache!.dividerThickness;
+        posLimitEnd += _sizesCache!.sizes[i + 1];
+        child2End += _sizesCache!.sizes[i];
+        child2End += _sizesCache!.dividerThickness;
+        child2End += _sizesCache!.sizes[i + 1];
+        posLimitEnd = math.max(
+            posLimitStart, posLimitEnd - _sizesCache!.minimalSizes[i + 1]);
+      }
+    }
+
+    _initialDrag = _InitialDrag(
+        initialDragPos: initialDragPos,
+        initialChild1Size: initialChild1Size,
+        initialChild2Size: initialChild2Size,
+        minimalChild1Size: minimalChild1Size,
+        minimalChild2Size: minimalChild2Size,
+        sumMinimals: sumMinimals,
+        sumSizes: sumSizes,
+        child1Start: child1Start,
+        child2End: child2End,
+        posLimitStart: posLimitStart,
+        posLimitEnd: posLimitEnd);
+    _initialDrag!.posBeforeMinimalChild1 = initialDragPos < posLimitStart;
+    _initialDrag!.posAfterMinimalChild2 = initialDragPos > posLimitEnd;
   }
 
   /// Calculates the new weights and sets if they are different from the current one.
   void _updateDifferentWeights(
       {required int childIndex,
       required double diffPos,
-      required double availableArea}) {
-    double minimalWeight1 =
-        _minimalWeight(childIndex: childIndex, availableArea: availableArea);
-    double minimalWeight2 = _minimalWeight(
-        childIndex: childIndex + 1, availableArea: availableArea);
-
-    double newChild1Weight =
-        ((_initialChild1Size + diffPos) * _initialChild1Weight) /
-            _initialChild1Size;
-
-    //double maxChild1Weight = _initialChild1Weight + _initialChild2Weight - minimalWeight1;
-
-    newChild1Weight = math.max(minimalWeight1, newChild1Weight);
-    //newChild1Weight = math.min(maxChild1Weight, newChild1Weight);
-
-    if (newChild1Weight < minimalWeight1) {
-      return;
-    }
-    double newChild2Weight =
-        _initialChild1Weight + _initialChild2Weight - newChild1Weight;
-
-    if (newChild2Weight < minimalWeight2) {
+      required double pos,
+      required double fullSize}) {
+    if (diffPos == 0) {
       return;
     }
 
-    if (_controller.getWeight(childIndex) != newChild1Weight &&
-        _controller.getWeight(childIndex + 1) != newChild2Weight) {
+    if (_initialDrag!.sumMinimals >= _initialDrag!.sumSizes) {
+      // minimals already smaller than available space. Ignoring...
+      return;
+    }
+
+    double newChild1Size;
+    double newChild2Size;
+
+    if (diffPos.isNegative) {
+      // divider moving on left/top from initial mouse position
+      if (_initialDrag!.posBeforeMinimalChild1) {
+        // can't shrink, already smaller than minimal
+        return;
+      }
+      newChild1Size = math.max(_initialDrag!.minimalChild1Size,
+          _initialDrag!.initialChild1Size + diffPos);
+      newChild2Size = _initialDrag!.sumSizes - newChild1Size;
+
+      if (_initialDrag!.posAfterMinimalChild2) {
+        if (newChild2Size > _initialDrag!.minimalChild2Size) {
+          _initialDrag!.posAfterMinimalChild2 = false;
+        }
+      } else if (newChild2Size < _initialDrag!.minimalChild2Size) {
+        double diff = _initialDrag!.minimalChild2Size - newChild2Size;
+        newChild2Size += diff;
+        newChild1Size -= diff;
+      }
+    } else {
+      // divider moving on right/bottom from initial mouse position
+      if (_initialDrag!.posAfterMinimalChild2) {
+        // can't shrink, already smaller than minimal
+        return;
+      }
+      newChild2Size = math.max(_initialDrag!.minimalChild2Size,
+          _initialDrag!.initialChild2Size - diffPos);
+      newChild1Size = _initialDrag!.sumSizes - newChild2Size;
+
+      if (_initialDrag!.posBeforeMinimalChild1) {
+        if (newChild1Size > _initialDrag!.minimalChild1Size) {
+          _initialDrag!.posBeforeMinimalChild1 = false;
+        }
+      } else if (newChild1Size < _initialDrag!.minimalChild1Size) {
+        double diff = _initialDrag!.minimalChild1Size - newChild1Size;
+        newChild1Size += diff;
+        newChild2Size -= diff;
+      }
+    }
+    if (newChild1Size >= 0 && newChild2Size >= 0) {
       setState(() {
-        _controller._setWeight(childIndex, newChild1Weight);
-        _controller._setWeight(childIndex + 1, newChild2Weight);
+        _sizesCache!.sizes[childIndex] = newChild1Size;
+        _sizesCache!.sizes[childIndex + 1] = newChild2Size;
       });
       if (widget.onSizeChange != null) {
         widget.onSizeChange!(childIndex, childIndex + 1);
@@ -620,27 +646,18 @@ class _MultiSplitViewState extends State<MultiSplitView> {
     return container.globalToLocal(globalPosition);
   }
 
-  double _distanceOnTheOppositeSide(
-      {required int childIndex,
-      required double totalChildrenSize,
-      required double totalRemainingWeight,
-      required MultiSplitViewThemeData themeData}) {
-    int amountRemainingDividers = widget.children.length - 1 - childIndex;
-    return (amountRemainingDividers * themeData.dividerThickness) +
-        (totalChildrenSize * totalRemainingWeight);
-  }
-
   /// Builds a [Positioned] using the [_DistanceFrom] parameters.
   Positioned _buildPositioned(
       {required _DistanceFrom distance,
       required Widget child,
       bool last = false}) {
-    return Positioned(
+    Positioned positioned = Positioned(
         top: _convert(distance.top, last),
         left: _convert(distance.left, last),
         right: _convert(distance.right, last),
         bottom: _convert(distance.bottom, last),
         child: child);
+    return positioned;
   }
 
   /// This is a workaround for https://github.com/flutter/flutter/issues/14288
@@ -665,6 +682,35 @@ class _DistanceFrom {
   double bottom;
 
   _DistanceFrom({this.top = 0, this.left = 0, this.right = 0, this.bottom = 0});
+}
+
+class _InitialDrag {
+  _InitialDrag(
+      {required this.initialDragPos,
+      required this.initialChild1Size,
+      required this.initialChild2Size,
+      required this.minimalChild1Size,
+      required this.minimalChild2Size,
+      required this.sumMinimals,
+      required this.sumSizes,
+      required this.child1Start,
+      required this.child2End,
+      required this.posLimitStart,
+      required this.posLimitEnd});
+
+  final double initialDragPos;
+  final double initialChild1Size;
+  final double initialChild2Size;
+  final double minimalChild1Size;
+  final double minimalChild2Size;
+  final double sumMinimals;
+  final double sumSizes;
+  final double child1Start;
+  final double child2End;
+  final double posLimitStart;
+  final double posLimitEnd;
+  bool posBeforeMinimalChild1 = false;
+  bool posAfterMinimalChild2 = false;
 }
 
 typedef OnSizeChange = void Function(int childIndex1, int childIndex2);